# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Система бронирования мест в кафе",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1/",
      "description": "Локальный сервер"
    }
  ],
...
  "paths": {
...
    "/cafes/{cafe_id}/time_slots": {
      "get": {
        "tags": [
          "Временные слоты"
        ],
        "summary": "Список временных слотов в кафе",
        "description": "Получение списка доступных для бронирования временных слотов в кафе. Для администраторов и менеджеров - все столы (с возможностью выбора), для пользователей - только активные.",
        "operationId": "get_all_time_slots_cafe__cafe_id__time_slots_get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            },
            "description": "ID кафе"
          },
          {
            "name": "show_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "title": "Показывать активные слоты?",
              "default": true
            },
            "description": "Показывать все слоты в кафе, только активные или только не активные. <br>По умолчанию показывает: <br><ul><li>для пользователя - только активные слоты в кафе (всегда, в не зависимости от значения параметра).</li><li>для администратора - все слоты в кафе (и активные и не активные)</li><li>для менеджера - активные слоты в кафе </li></ul>"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimeSlotInfo"
                  },
                  "title": "Возвращает все слоты в кафе"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Временные слоты"
        ],
        "summary": "Новый временной слот в кафе",
        "description": "Создает нового временного слота в кафе. Только для администраторов и менеджеров.",
        "operationId": "create_time_slot_cafe__cafe_id__time_slots_post",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            },
            "description": "ID кафе"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeSlotCreate",
                "description": "Данные нового временного слота"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSlotInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    },
    "/cafes/{cafe_id}/time_slots/{slot_id}": {
      "get": {
        "tags": [
          "Временные слоты"
        ],
        "summary": "Информация о временном слоте в кафе по его ID",
        "description": "Получение информации о временном слоте в кафе по его ID. Для администраторов и менеджеров - все столы, для пользователей - только активные.",
        "operationId": "get_time_slot_by_id_cafe__cafe_id__time_slots__slot_id__get",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            },
            "description": "ID кафе"
          },
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ID слота",
              "title": "Slot Id"
            },
            "description": "ID слота"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSlotInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Временные слоты"
        ],
        "summary": "Обновление информации о временом слоте в кафе по его ID",
        "description": "Обновление информации о временом слоте в кафе по его ID. Только для администраторов и менеджеров.",
        "operationId": "path_time_slot_by_id_cafe__cafe_id__time_slots__slot_id__patch",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cafe Id"
            },
            "description": "ID кафе"
          },
          {
            "name": "slot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "ID слота",
              "title": "Slot Id"
            },
            "description": "ID слота"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimeSlotUpdate",
                "description": "Обновленные данные временного слота"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешно",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimeSlotInfo"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка в параметрах запроса",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "401": {
            "description": "Неавторизированный пользователь",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "403": {
            "description": "Доступ запрещен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "404": {
            "description": "Данные не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          },
          "422": {
            "description": "Ошибка валидации данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CafeShortInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "photo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Photo Id"
          }
        },
        "type": "object",
        "title": "CafeShortInfo"
      },
...
      "CustomError": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "CustomError"
      },
...
      "TimeSlotCreate": {
        "properties": {
          "start_time": {
            "type": "string",
            "format": "time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "time",
            "title": "End Time"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "start_time",
          "end_time"
        ],
        "title": "TimeSlotCreate"
      },
      "TimeSlotInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "cafe": {
            "$ref": "#/components/schemas/CafeShortInfo"
          },
          "start_time": {
            "type": "string",
            "format": "time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "time",
            "title": "End Time"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "TimeSlotInfo"
      },
...
      "TimeSlotUpdate": {
        "properties": {
          "start_time": {
            "type": "string",
            "format": "time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "time",
            "title": "End Time"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "TimeSlotUpdate"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}